+++ src/models/task.h
@@ -9,7 +9,7 @@ class Task: public QObject
 {
     Q_OBJECT
 
-    Q_PROPERTY(int id READ getId CONSTANT)
+    Q_PROPERTY(long long id READ getId CONSTANT)
     Q_PROPERTY(bool approvalStatus READ getApprovalStatus WRITE setApprovalStatus NOTIFY approvalStatusChanged)
     Q_PROPERTY(QString globalId READ getGlobalId WRITE setGlobalId NOTIFY globalIdChanged)
     Q_PROPERTY(QString taskName READ getTaskName WRITE setTaskName NOTIFY taskNameChanged)
@@ -17,7 +17,7 @@ class Task: public QObject
     Q_PROPERTY(QString bimElement READ getBimElement WRITE setBimElement NOTIFY bimElementChanged)
     Q_PROPERTY(QString startDate READ getStartDate WRITE setStartDate NOTIFY startDateChanged)
     Q_PROPERTY(QString endDate READ getEndDate WRITE setEndDate NOTIFY endDateChanged)
-    Q_PROPERTY(int pid READ getParentId WRITE setParentId NOTIFY parentIdChanged)
+    Q_PROPERTY(long long pid READ getParentId WRITE setParentId NOTIFY parentIdChanged)
 
     Q_PROPERTY(int days READ getDuration WRITE setDuration NOTIFY durationChanged)
 
@@ -57,16 +57,16 @@ class Task: public QObject
 
 public:
     explicit Task(QObject* parent = nullptr): QObject(parent) {}
-    Task(int id, const QString& globalId, bool approvalStatus,
+    Task(long long id, const QString& globalId, bool approvalStatus,
              const QString& taskName,
              const QString& description,
              const QString& bimElement,
              const QString& startDate,
              const QString& endDate,
-             int pid,
+             long long pid,
             QObject* parent = nullptr);
 
-    int getId() const { return id; }
+    long long getId() const { return id; }
     QString getGlobalId() const { return globalId; }
     bool getApprovalStatus() const { return approvalStatus; }
     QString getTaskName() const { return taskName; }
@@ -74,7 +74,7 @@ public:
     QString getBimElement() const { return bimElement; }
     QString getStartDate() const { return startDate; }
     QString getEndDate() const { return endDate; }
-    int getParentId() const { return pid; }
+    long long getParentId() const { return pid; }
     int getDuration() const { return days; }
     int getYear() const { return year; }
     int getMonth() const { return month; }
@@ -107,7 +107,7 @@ public:
     QString getMonth_Year() const { return month_year; }
 
 
-    void setId(int id) { this->id = id; }
+    void setId(long long id) { this->id = id; }
     void setGlobalId(const QString& globalId) { this->globalId = globalId; }
     void setApprovalStatus(bool status) { this->approvalStatus = status; }
     void setTaskName(const QString& taskName) { this->taskName = taskName; }
@@ -115,7 +115,7 @@ public:
     void setBimElement(const QString& bimElement) { this->bimElement = bimElement; }
     void setStartDate(const QString& startDate) { this->startDate = startDate; }
     void setEndDate(const QString& endDate) { this->endDate = endDate; }
-    void setParentId(int pid) { this->pid = pid; }
+    void setParentId(long long pid) { this->pid = pid; }
     void setDuration(const int days) { this->days = days; }
     void setYear(const int year) { this->year = year; }
     void setMonth(const int month) { this->month = month; }
@@ -191,7 +191,7 @@ signals:
     void month_yearChanged();
 
 private:
-    int id = 0;
+    long long id = 0;
     QString globalId;
     bool approvalStatus = true;
     QString taskName;
@@ -200,7 +200,7 @@ private:
     QString startDate;
     QString endDate;
     int days = 0;
-    int pid = 0;
+    long long pid = 0;
 
     int year = 0;
     int month = 0;
